Expedia-Hotel-Recommendations
Expedia gives you customer behaviour, including what customers search for, they and the search results do not represent this competition (whether to click to read), and the data for the search results for travel packages is a temporary selection from Expedia. Statistical data.
Expedia is expected to form the home algorithmic hotel group that users book. The price hotel group within the city will be searched for similar hotels (based on history, customer star ratings, categories of related centres, etc.) together. These groups of these hotels can attentively identify the types of hotels that people are going to book, while new hotels for which there is no historical data.
Our goal is to predict attribute results for user events (event user base) based on the user's searches and other content related to that user's subscriptions.